Admin API now uses latest version of restlet

+++ platform/org.ossmeter.platform.admin/src/org/ossmeter/platform/admin/FullMetricAnalysis.java
@@ -11,8 +11,9 @@
 package org.ossmeter.platform.admin;
 
 import java.net.UnknownHostException;
+
 import org.ossmeter.platform.Configuration;
-import org.restlet.engine.header.Header;
+import org.restlet.data.Header;
 import org.restlet.resource.Get;
 import org.restlet.resource.ServerResource;
 import org.restlet.util.Series;
@@ -65,7 +66,7 @@ public class FullMetricAnalysis extends ServerResource {
 						on.put("millis", (Double)obj.get("millisTaken"));
 						results.add(on);
 					} else {
-						results.add((Double)obj.get("millisTaken"));
+						results.add((Integer)obj.get("millisTaken"));
 					}
 				}
